MVS: multi-state with canvas node
